requirements.txt
requests==2.31.0
python-dotenv==1.0.0
tweepy==4.14.0
openai==1.3.0
praw==7.7.1
facebook-sdk==3.1.0
linkedin-api==2.0.0 